error fix

+++ compute/src/main/scala/com/ums/HashMapAccumalator.scala
@@ -11,6 +11,6 @@ object HashMapAccumalatorParam extends AccumulatorParam[Map[String, Double]]{
   }
 
   override def addInPlace(r1: Map[String, Double], r2: Map[String, Double]):Map[String, Double]= {
-    r1 ++ r2.map{ case (k, v) => k -> (v + r1.getOrElse(k, 0))}
+    r1 ++ r2.map{ case (k, v) => k -> (v + r1.getOrElse(k, 0.toDouble))}
   }
 }
